pipes added

+++ src/hushlib.h
@@ -9,7 +9,7 @@ void init();
 
 int8_t* greeting();
 
-int8_t handle_command(int8_t** args);
+int32_t handle_command(int8_t** args);
 
 int8_t** parse_args(int8_t* input_line);
